Ensure the VSS ready messages are signed
The DKG election requires signed proof enough distinct readies were
obtained from a VSS for it to be able to complete. These signed readies
are distributed from the leader to the follower nodes so that the
leader's claim of which VSSes it wants to use for the final key can be
verified.